Adds core mobule and moves submodules
