+++ src/core.rs
@@ -1,0 +1,7 @@
+// Copyright 2024 Dotanuki Labs
+// SPDX-License-Identifier: MIT
+
+mod disk;
+pub mod models;
+mod ram;
+pub mod wiper;
