+++ src/core/disk.rs
