+++ src/core/ram/jps.rs
